Create README.md
Designed as a web-based solution, the Online Doctor Appointment and Medicine System offers a user-friendly platform that streamlines the process of scheduling appointments, ordering medications, and accessing medical services from the comfort of one's home. With a focus on user convenience and privacy, this system simplifies healthcare management for both doctors and patients. By incorporating an intuitive interface and automated reminders, it ensures a seamless experience for users. This platform enhances efficiency for doctors, enabling them to manage appointments and medical services effectively. Similarly, patients benefit from the convenience of scheduling appointments and ordering medications online, saving time and effort. By prioritizing patient privacy and data security, this system fosters trust and confidence in the digital healthcare experience.